4. inner class를 상위 클래스로 이동
