+++ src/test/java/com/adventours/calendar/KakaoOAuthRequestAdapter.java
@@ -1,0 +1,8 @@
+package com.adventours.calendar;
+
+public class KakaoOAuthRequestAdapter implements OAuthRequestPort {
+    @Override
+    public OAuthUserInformation requestUserInformation(final String token) {
+        throw new UnsupportedOperationException("KakaoOAuthRequestAdapter#requestUserInformation not implemented yet.")
+    }
+}
