+++ src/test/java/com/adventours/calendar/LoginRequest.java
@@ -1,0 +1,4 @@
+package com.adventours.calendar;
+
+record LoginRequest(String token) {
+}
