+++ src/test/java/com/adventours/calendar/LoginResponse.java
@@ -1,0 +1,4 @@
+package com.adventours.calendar;
+
+record LoginResponse(long memberId, boolean isNewMember) {
+}
