+++ src/test/java/com/adventours/calendar/Member.java
@@ -1,0 +1,9 @@
+package com.adventours.calendar;
+
+class Member {
+    private Long id;
+
+    public Long getId() {
+        return id;
+    }
+}
