+++ src/test/java/com/adventours/calendar/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.adventours.calendar;
+
+import java.util.Optional;
+
+interface MemberRepository {
+    Optional<Member> findByProviderAndProviderId(OAuthProvider provider, String s);
+
+    Member save(Member member);
+}
