+++ src/test/java/com/adventours/calendar/OAuthProvider.java
@@ -1,0 +1,5 @@
+package com.adventours.calendar;
+
+enum OAuthProvider {
+    KAKAO
+}
