+++ src/test/java/com/adventours/calendar/OAuthRequestPort.java
@@ -1,0 +1,5 @@
+package com.adventours.calendar;
+
+interface OAuthRequestPort {
+    OAuthUserInformation requestUserInformation(String token);
+}
