+++ src/test/java/com/adventours/calendar/OAuthUserInformation.java
@@ -1,0 +1,4 @@
+package com.adventours.calendar;
+
+record OAuthUserInformation(String providerId, String initialProfileImage) {
+}
